requirements.txt
pytorch_lightning==1.9.0
opencv-python
tqdm
PIL
numpy
kornia
webdataset
omegaconf
einops
scikit-image
pymcubes
carvekit-colab
open3d
easydict
nerfacc
git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch
git+https://github.com/openai/CLIP.git